requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pyyaml==6.0.1
typer==0.9.0
pytest==7.4.3
httpx==0.25.2
pandas==2.1.3
pyarrow==14.0.1
